DMP-54 Add Notification table (#52)
* DMP-54 Add Notification table
exclude dto and entity classes from code coverage
