+++ src/integrationTest/java/uk/gov/hmcts/darts/common/controller/DailyListTest.java
@@ -4,10 +4,12 @@ import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
+import uk.gov.hmcts.darts.notification.repository.NotificationRepository;
 
 import java.io.File;
 import java.io.IOException;
@@ -19,6 +21,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @WebMvcTest
 class DailyListTest {
+    @MockBean
+    private NotificationRepository notificationRepository;
 
     @Autowired
     private transient MockMvc mockMvc;
