+++ src/main/java/uk/gov/hmcts/darts/common/entity/Annotation.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.common.entities;
+package uk.gov.hmcts.darts.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,8 +6,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
@@ -15,50 +14,30 @@ import jakarta.persistence.Version;
 import lombok.Data;
 
 import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
 
 @Entity
-@Table(name = "moj_transcription")
+@Table(name = "moj_annotation")
 @Data
-public class Transcription {
+public class Annotation {
 
     @Id
-    @Column(name = "moj_tra_id")
+    @Column(name = "moj_ann_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @ManyToOne
+    @OneToOne
     @JoinColumn(name = "moj_cas_id")
     private Case theCase;
 
-    @Column(name = "r_transcription_object_id", length = 16)
+    @Column(name = "r_annotation_object_id", length = 16)
     private String legacyObjectId;
 
-    @Column(name = "c_company", length = 64)
-    private String company;
+    @Column(name = "c_text", length = 2000)
+    private String text;
 
-    @Column(name = "c_type")
-    private Integer type;
-
-    @Column(name = "c_notification_type", length = 64)
-    private String notificationType;
-
-    @Column(name = "c_urgent")
-    private Integer urgent;
-
-    @Column(name = "c_requestor", length = 32)
-    private String requestor;
-
-    @Column(name = "c_current_state", length = 32)
-    private String currentState;
-
-    @Column(name = "c_urgency", length = 32)
-    private String urgency;
-
-    @Temporal(TemporalType.DATE)
-    @Column(name = "c_hearing_date")
-    private Date hearingDate;
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "c_time_stamp")
+    private Date timestamp;
 
     @Temporal(TemporalType.DATE)
     @Column(name = "c_start")
@@ -89,7 +68,4 @@ public class Transcription {
     @Version
     @Column(name = "i_version_label")
     private Short version;
-
-    @OneToMany(mappedBy = "theTranscription")
-    private Set<TranscriptionComment> theTranscriptionComments = new HashSet<>();
 }
