+++ src/main/java/uk/gov/hmcts/darts/notification/dto/CreateNotificationReq.java
@@ -1,0 +1,11 @@
+package uk.gov.hmcts.darts.notification.dto;
+
+import lombok.Data;
+
+@Data
+public class CreateNotificationReq {
+    String eventId;
+    String caseId;
+    String emailAddress;
+    String templateValues;
+}
