+++ src/main/java/uk/gov/hmcts/darts/notification/enums/NotificationStatus.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.darts.notification.enums;
+
+public enum NotificationStatus {
+    OPEN,
+    PROCESSING,
+    FAILED,
+    SENT;
+}
