+++ src/main/java/uk/gov/hmcts/darts/notification/service/NotificationService.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.darts.notification.service;
+
+import uk.gov.hmcts.darts.notification.dto.CreateNotificationReq;
+import uk.gov.hmcts.darts.notification.entity.Notification;
+
+public interface NotificationService {
+    Notification sendNotification(CreateNotificationReq createNotificationReq);
+}
